Initial commit of the Project.
Ok guys, this time is the One. I _will_ finish this project.

So let me explain it : I'm huge (like really HUGE) nerd, which have
the dream to live in all the sci-fi/fantasy world he read/saw. Then
one time, I found that book in an old refurb store : Player One by
Ernest Cline (Waaayyyyyyy before it became the Spielberg's movie).

It was like a revelation. Virtual Reality, living in real nerd
worlds ! I thought : Ok I want this. But it seems that nobody is
interested in that.

So f*** it, I'm a programmer, I'll do it myself. I don't care if it
sucks for people, or if it take me 10 years, I **need** this.

The main purpose will not to _sell_ this, but just to enjoy it myself
(sounds better in my head). Maybe I'll extract a reusable engine out
of it and make original games with it, but maybe not.

We have the capacity to create hundred and hundred of worlds right at
the end of our finger tips, using computing processing that people
who sends men on the moon only dreamt about. And we keep making the
same shity forms, or bullshit ads systems and algorithm's social
look-how-i-destroy-my-life-for-celebrity network. Computers and
Internet gives us access to all the available knowledge in the history,
I can't stand looking how we misuse it.

I want to build something that matters.

At least for me.