Windows Runtime support is on track.
That's cool (in all humility), that's damn cool.

I've implemented the Windows Runtime specific code (to handle a
Window creation), and I managed to do it without touching my engine.
I've added a brand new module (a separate VS project) to my solution
which use Engine/includes/platforms as includes, and then I implement
the winrt.hpp header in this module. After that, all I have to do, is
to link this new module to my Demo executable and baaam magic is done.
Now I can only change a simple #define in my Demo main.cpp to switch
between platforms. And the bonus part ? Given the fact that there's no
coupling between the module and the engine (beside the header), when
I change the #define in my Demo, it only rebuild the Demo project, not
all the solution. Fast as hell.

I remember that I didn't answer a question that's maybe you've been
asking yourself : Why the f*** did I start this project in C ?

Because I LOVE C. I mean C is a complete language, with a specification
and standard library which can be contains on a postcard, low level
mechanisms, simplicity (I mean no public/private, no template, no class
mechanisms to respect and implement...), and completely gives you the
control on what you build.

I know that most of the languages now implements security (like a memory
security). But honestly, sometime we may want to do something which is
considered as a bad code to read, but which is fast. And THAT'S
what matters for me : performance.

With C you can strictly decide how your memory bits will be arranged, how
many CPU cycles will your algorithms takes, you have control over
everything. Which means that you can fine tune your apps to really use
the full potential of a computer.

Nowdays computers have the compute power to move mountains, and still
99% of the apps on the market are as slow as sleep slugs in a blackhole
accretion disk ! Because we want to help developers create those apps.
But as a developer let me tells it : WE DON'T CARE ABOUT DEVELOPERS. We
care about users ! Do you want to eat garbage, just because it will help
the 5 Stars Chef ? No. Do you want to leave on an house with poor foundations
and walls, just because it will help the architect ? No. Same thing applies
to programs. And in my opinion, C still is the only language which can
perfectly guaranties that.
(If this subject interest you, I strongly recommand that you take a look
on https://handmade.network : a community of developers thinking the same
way and making blazing fast applications.)

Another reason why I use it, and the most important :

It's my side project. I love C. I don't plan on building a team and selling
it (yet). So I do what I want as long as I have fun.