It's been a while since my last entry...
It was difficult to find some time for working on this project (I changed my
daily job and was still on probation period, family time, little games
too -- still not finished Hogwarts Legacy). The little amount I found, I spent
it learning Vulkan and trying to use it in the project.

I have some previous experience with DirectX 12 (from the past iterations of
this project), and Vulkan is quite similar in the approch. But bloody hell that
this API is verbose... It's at least 30LoC to create just one resource and we
will need at least Instance, Device (Physical&logical), Commands
Queue/Allocator/Buffer, Swapchain...

But after some time, I thought that I've shifted from my initial goal with this
project : Keep It Simple. Vulkan has a multiplatform support, which means that
specific code (like handling HWND or HINSTANCE on Windows) must be done through
extensions to enable, more specific headers to includes, and .lib/.dll to add.
And it's not the simplest way to do low-level graphic programming on Windows,
because it will require installation of external tools and libraries as long as
a lot of checks to make sure that  the platform supports the extensions.

On the other end, Windows already have a low-level graphic API : DirectX12.
Headers and libs are already bundled in all Visual C++ redistribuable packages,
and there is no need to load runtime extensions because we _know_ the platform
it works on.

So I decided to put on hold the vulkan implementation and go back to a DirectX12
one instead. I've just backport the core mechanics around the game level
management.