Reworking the project management.
I decided to take a little time to properly restructure the project.
All the modules are supposed to be independant, but they're all in the
same repository. This can be dangerous, because by default I'll have
access to the full API, which might confuse the independancy thing.

So I decided to create 3 different repositories (and GitHub repos) :
two publics (the Engine here, and the WinRT module), and one private
which is my personal demo of the engine.

Next step I'll setup some automated actions (build, validation,
packaging...), and leverage a lot GitHub tools (Issues, Projects)
to setup a complete roadmap.

Once again, the purpose of this Engine is :
1. to learn how to build a low level game engine from scratch
2. to create the game I've always dreamt of, only _for me_. No sale.

So this is this game that will drive the evolutions of my engine, like
StarCitizen/Squadron42 does.

Yes, I've quoted this. Fight me.