DirectX 12 is now initialized in the project.
I've spent a little time doing some project management over the engine. Why ?
Because I'm currently at the 6th iteration of it, because everytime it turns
out I go in all the direction at once.

Now, I've set up a roadmap (private on my GitHub account, because it also
tracks the progress of the poc game), with the hard constraints of finishing
the first version of the game by September, the 1st.

Funny enough, I've already overburn the estimation I've made for Dx12 init
implementation (life, family...), so I had to reduce the time I've allowed
to map the game levels.

I've also separated the project in several repositories : this one for the
engine itself, and one repository per module implementation. Which for now
gives me two repositories:  WinRT platform and Dx12 renderer.